pb_type.xml generation basics implemented
