+++ utils/vlog/.gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.pyc
 dump.json
 pb_type.xml
+merged.xml
